feat: commented utils
