+++ src/system/information/stores/OutputStore.ts
@@ -1,15 +1,15 @@
-import { Time } from "../../shared/time";
-import { getSubFiles, getSubFolders } from "../util";
+import { Time } from "../../../shared/time";
+import { getSubFiles, getSubFolders } from "../../util";
 import { AbstractStore } from "./AbstractStore";
-import { CSVRebateFile } from "./items/CSVRebateFile";
+import { ExcelRebateFile } from "../items/ExcelRebateFile";
 
 /** ------------------------------------------------------------------------- */
 
-type Item = CSVRebateFile<{ quarter: Time }>;
+type Item = ExcelRebateFile<{ quarter: Time }>;
 interface Meta { directory: string };
 
-export class TruthStore extends AbstractStore<Item, Meta> {
-  public readonly name = "truths";
+export class OutputStore extends AbstractStore<Item, Meta> {
+  public readonly name = "outputs";
 
   public async gather(): Promise<void> {
     for (const [time_path, time_str] of await getSubFolders(this.meta.directory)) {
@@ -17,7 +17,7 @@ export class TruthStore extends AbstractStore<Item, Meta> {
       if (time == null) continue;
 
       for (const [file_path] of await getSubFiles(time_path)) {
-        this.add(new CSVRebateFile(file_path, { quarter: time }));
+        this.add(new ExcelRebateFile(file_path, { quarter: time }));
       }
     }
   }
