+++ src/system/information/stores/ReferenceStore.ts
@@ -1,7 +1,7 @@
 import path from "path";
 import { AbstractStore } from "./AbstractStore";
-import { getSubFiles } from "../util";
-import { ReferenceFile } from "./items/ReferenceFile";
+import { getSubFiles } from "../../util";
+import { ReferenceFile } from "../items/ReferenceFile";
 
 /** ------------------------------------------------------------------------- */
