+++ src/system/information/stores/UtilityStore.ts
@@ -1,8 +1,8 @@
 import path from "path";
 import { AbstractStore } from "./AbstractStore";
-import { getSubFiles, getSubFolders } from "../util";
-import { ReferenceFile } from "./items/ReferenceFile";
-import { Time } from "../../shared/time";
+import { getSubFiles, getSubFolders } from "../../util";
+import { ReferenceFile } from "../items/ReferenceFile";
+import { Time } from "../../../shared/time";
 
 /** ------------------------------------------------------------------------- */
