+++ src/system/util.ts
@@ -2,19 +2,32 @@ import { z } from "zod/v4";
 import { readdir } from 'fs/promises';
 import path from "path";
 import { Rebate } from "../shared/worker/response";
-import moment from "moment";
 
 /** ------------------------------------------------------------------------- */
 
+/**
+ * A schema that automatically converts any numbers, strings as numbers, or
+ * Excel indices (A, B, ...) to 0-based indices.
+ */
 export const ExcelIndexSchema = z.union([
   z.number(),
   z.string().regex(/[A-Z]+/)
 ]).transform(s => getTrueIndex(s));
 
+/**
+ * Convert an Excel index into a 0-based index.
+ * @param letters The index to parse. Must be valid.
+ * @returns The 0-based index.
+ */
 function getIndexFromExcel(letters: string): number {
   return letters.split("").reduce((s, c) => c.charCodeAt(0) - 64 + s * 26, 0) - 1;
 }
 
+/**
+ * Turn a number into an Excel index.
+ * @param n The number.
+ * @returns An Excel index.
+ */
 export function getExcelFromIndex(n: number) {
   n++;
   let res = "";
@@ -34,6 +47,11 @@ export function getExcelFromIndex(n: number) {
   return res.split("").reverse().join("");
 }
 
+/**
+ * Attempts to index-like a value into a number.
+ * @param index An index-like value. Could be an Excel index, or just a normal number.
+ * @returns A number.
+ */
 export function getTrueIndex(index: string | number): number {
   if (typeof index === "number") return index;
   return getIndexFromExcel(index);
@@ -41,24 +59,22 @@ export function getTrueIndex(index: string | number): number {
 
 /** ------------------------------------------------------------------------- */
 
+/**
+ * Create hash unique hash for a rebate.
+ * @param rebate The rebate.
+ * @returns The unique hash.
+ */
 export function getRebateHash(rebate: Rebate): string {
   const { transactionDate, supplierId, memberId, distributorName, purchaseAmount, rebateAmount, invoiceId, invoiceDate } = rebate;
   return `${transactionDate},${supplierId},${memberId},${distributorName},${purchaseAmount},${rebateAmount},${invoiceId},${invoiceDate}`;
 }
 
-const COMMON_PARSE = ["MM/DD/YY", "M/D/YYYY"];
-
-export function areRebatesEqual(a: Rebate, b: Rebate) {
-  return a.invoiceId === b.invoiceId
-    && Math.abs(a.purchaseAmount - b.purchaseAmount) <= 0.02
-    && Math.abs(a.rebateAmount - b.rebateAmount) <= 0.02
-    && moment(a.invoiceDate, COMMON_PARSE).format("MM/DD/YY") === moment(b.invoiceDate, COMMON_PARSE).format("MM/DD/YY")
-    && moment(a.transactionDate, COMMON_PARSE).format("MM/DD/YY") === moment(b.transactionDate, COMMON_PARSE).format("MM/DD/YY")
-    && a.supplierId === b.supplierId
-    && a.memberId === b.memberId
-    // && a.distributorName === b.distributorName;
-}
-
+/**
+ * Partition a set of objects into seperate buckets, based upon the value of a certain property.
+ * @param objects The set of objects.
+ * @param key The proerty to split on.
+ * @returns A list of buckets, where each bucket contains a set of objects where their `key` is equal.
+ */
 export function getPartition<O extends object, K extends keyof O>(objects: O[], key: K): Map<O[K], O[]> {
   const buckets = new Map<O[K], O[]>();
   for (const object of objects) {
@@ -73,56 +89,22 @@ export function getPartition<O extends object, K extends keyof O>(objects: O[], 
   return buckets;
 }
 
-export class RebateSet {
-  private buckets: Record<string, Rebate[]>;
-
-  constructor(rebates: Rebate[]) {
-    this.buckets = {};
-
-    for (const rebate of rebates) {
-      this.give(rebate);
-    }
-  }
-
-  public values() {
-    return Object.values(this.buckets).flat();
-  }
-
-  public find(rebate: Rebate): Maybe<[string, number]> {
-    const bucket = this.buckets[rebate.invoiceId];
-    if (bucket == null) return null;
-
-    for (let i = 0; i < bucket.length; i++) {
-      if (areRebatesEqual(bucket[i], rebate)) {
-        return [rebate.invoiceId, i];
-      }
-    }
-
-    return null;
-  }
-
-  public has(rebate: Rebate) {
-    return this.find(rebate) != null;
-  }
-
-  public give(rebate: Rebate) {
-    (this.buckets[rebate.invoiceId] ??= []).push(rebate);
-  }
-
-  public take(rebate: Rebate) {
-    const place = this.find(rebate);
-    if (place == null) return false;
-
-    this.buckets[place[0]].splice(place[1], 1);
-    return true;
-  }
-}
-
+/**
+ * Create a copy of a table, where all of the rows point back to this new table.
+ * @param table The tabe to copy.
+ * @returns A new table. The rows are not duplicated.
+ */
 export function rewire(table: Table) {
   table.data.forEach(r => { r.table = table; });
   return table;
 }
 
+/**
+ * Create a new table, based on a matrix of data.
+ * @param rows The data.
+ * @param path The file path of the table, if required.
+ * @returns A new table.
+ */
 export function makeTable(rows: string[][], path = "") {
   const table: Table = { data: [], path };
   table.data = rows.map(r => ({ data: r, table }));
@@ -130,38 +112,22 @@ export function makeTable(rows: string[][], path = "") {
   return table;
 }
 
+/**
+ * Given a directory, find all files in the directory. This is recursive.
+ * @param directory The directory to search.
+ * @returns All files found. Each file is a list containing (1) its absolute path, and (2) its name.
+ */
 export async function getSubFiles(directory: string): Promise<[string, string][]> {
   const entries = await readdir(directory, { withFileTypes: true, recursive: true });
   return entries.filter(e => e.isFile()).map(e => [path.join(e.parentPath, e.name), e.name]);
 }
 
+/**
+ * Given a directory, find all folders in the directory. This is not recursive.
+ * @param directory The directory to search.
+ * @returns All folders found. Each folder is a list containing (1) its absolute path, and (2) its name.
+ */
 export async function getSubFolders(directory: string): Promise<[string, string][]> {
   const entries = await readdir(directory, { withFileTypes: true });
   return entries.filter(e => e.isDirectory()).map(e => [path.join(e.parentPath, e.name), e.name]);
 }
-
-/** ------------------------------------------------------------------------- */
-
-// function getUnitCubicCurve(x: number, angle: number) {
-//   return (angle - 2) * (x ** 3) + (-2 * angle + 3) * (x ** 2) + angle * x;
-// }
-
-// function getCubicCurve(x: number, angle: number, start: [number, number], end: [number, number]) {
-//   const [a, b] = start;
-//   const [c, d] = end;
-
-//   return (d - b) * getUnitCubicCurve((x - a) / (c - a), angle * (c - a) / (d - b)) + b
-// }
-
-// export function useFancyAnimation(value: number, delay: number, handler: (value: number) => void) {
-//   const previous = usePrevious(value);
-
-//   React.useEffect(() => {
-//     let timer = delay;
-//     const token = setInterval(() => {
-//       if (timer < delay) return clearInterval(token);
-//       timer -= 100;
-      
-//     }, 100);
-//   }, [value, delay]);
-// }
